requirements.txt
pytest==8.0.0
selene==2.0.0rc4
allure-pytest
curlify~=2.2.1
requests~=2.31.0